+++ src/types.ts
@@ -278,3 +278,24 @@ export interface IModelOptions {
     automaticName?: boolean;
   };
 }
+
+/** This Enum is meant for baseProp to decide for diffrent props (like if it is an arrayProp or prop or mapProp) */
+export enum WhatIsIt {
+  ARRAY,
+  MAP,
+  NONE
+}
+
+export interface DecoratedPropertyMetadata {
+  /** Prop Options */
+  origOptions: any;
+  /** What the Property Type should be */
+  Type: AnyParamConstructor<any>;
+  /** Target Class */
+  target: any;
+  /** Property name */
+  key: string;
+  /** What is it for a prop type? */
+  whatis: WhatIsIt;
+}
+export type DecoratedPropertyMetadataMap = Map<string, DecoratedPropertyMetadata>;
